% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2pairwisemat.R
\name{df2pmat}
\alias{df2pmat}
\title{Dataframe to pairwise matrix converter}
\usage{
df2pmat(df, var, grp1, grp2)
}
\arguments{
\item{df}{A long-form dataframe containing two columns of grouping variables and a column of values.}

\item{var}{Column name containing the value of interest (to populate the matrix).}

\item{grp1}{Grouping variable column 1 (e.g., population).}

\item{grp2}{Grouping variable column 2 (e.g., population).}
}
\value{
Distance matrix of the class Matrix.
}
\description{
Converts a long-form dataframe containing pairwise values into a symmetrical matrix.
}
