% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insettr.R
\name{insettr}
\alias{insettr}
\title{Inset plots with ggplot2}
\usage{
insettr(plot1, plot2, location, h, w)
}
\arguments{
\item{plot1}{The main ggplot object.}

\item{plot2}{A ggplot object to be inset.}

\item{location}{The corner the inset plot is assigned to (must be one of "bl", "br", "tl", "tr").}

\item{h}{Height of the inset plot (as a \% of the main plot size).}

\item{w}{Width of the inset plot (as a \% of the main plot size).}
}
\value{
A ggplot object.
}
\description{
This function allows you to specify the corner the inset plot should be
positioned in, rather than requiring the user supply specific x and y coordinates.
}
