% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_map.R
\name{map}
\alias{map}
\title{Custom map of BC and surrounding area}
\usage{
map(xmin, xmax, ymin, ymax, water_colour, na_inset)
}
\arguments{
\item{xmin}{Minimum longitude.}

\item{xmax}{Maximum longitude.}

\item{ymin}{Minimum latitude.}

\item{ymax}{Maximum latitude.}
}
\description{
Custom map of BC and surrounding area
}
