% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcftools_freq.R
\name{read_frq}
\alias{read_frq}
\title{Read *.frq file from VCFtools}
\usage{
read_frq(file)
}
\arguments{
\item{file}{File path to a *.frq file produced by VCFtools. Column labels assume that files were generated with the flag --derived.}
}
\value{
Tidy dataframe.
}
\description{
A function for reading in and formatting the output of VCFtools --freq --derived into a tidy data frame.
}
